requirements.txt
django==1.7.4
django-ckeditor==4.4.4
